+++ lib/fluffy/backend/contract.ex
@@ -7,6 +7,8 @@ defmodule Fluffy.Backend.Contract do
 
   @callback start_session(keyword(), TestScope.attachment()) :: Session.t()
   @callback close_session(Session.t()) :: :ok | {:error, term()}
+  @callback new_page(Session.t(), term()) :: Session.t()
+  @callback history(Session.t(), :go_back | :go_forward, keyword()) :: Session.t()
   @callback visit(Session.t(), String.t()) :: Session.t()
   @callback reload(Session.t(), keyword()) :: Session.t()
   @callback navigate(Session.t(), Navigation.t()) :: Session.t()